created custome hooks for IconText, PostFooterIconList, refactored code for UserPostBottom, and UserPos
