+++ frontend/components/molecules/IconText/IconText.jsx
@@ -1,5 +1,6 @@
 import Image from "next/image";
 import icons from "@/constants/icons";
+import useIconText from "./useIconText";
 export default function IconText({
   src,
   width,
@@ -10,57 +11,17 @@ export default function IconText({
   user_has_liked = false,
   like_type = null,
 }) {
-  const getLikeTypeColor = () => {
-    if (!user_has_liked) {
-      return "text-slate-600";
-    }
-    switch (like_type) {
-      case "Like":
-        return "text-blue-600";
-      case "Clap":
-        return "text-green-600";
-      case "Support":
-        return "text-purple-600";
-      case "Heart":
-        return "text-red-600";
-      case "Bulb":
-        return "text-yellow-600";
-      case "Laugh":
-        return "text-sky-600";
-      default:
-        return "text-slate-600";
-    }
-  };
-  const getIconSource = (src, user_has_liked, like_type) => {
-    if (user_has_liked) {
-      const icon = icons.find((icon) => icon.alt === like_type);
-      return icon ? icon.src : src;
-    }
-    return src;
-  };
-
-  const getGroupHoverTextColor = () => {
-    if (!user_has_liked) {
-      return "text-slate-900";
-    }
-    switch (like_type) {
-      case "Like":
-        return "text-blue-900";
-      case "Clap":
-        return "text-green-900";
-      case "Support":
-        return "text-purple-900";
-      case "Heart":
-        return "text-red-900";
-      case "Bulb":
-        return "text-yellow-900";
-      case "Laugh":
-        return "text-sky-900";
-      default:
-        return "text-slate-900";
-    }
-  };
-  const data = `w-15 h-15`;
+  const { getLikeTypeColor, getIconSource, getGroupHoverTextColor, data } =
+    useIconText(
+      src,
+      width,
+      height,
+      alt,
+      text,
+      onClick,
+      user_has_liked,
+      like_type
+    );
   return (
     <div
       className="flex flex-1 p-2 my-3 mx-1 space-x-1 justify-center items-center rounded-md hover:bg-slate-100 hover:cursor-pointer"
