+++ frontend/components/molecules/UserPost/UserPostBottom/UserPostBottom.jsx
@@ -3,19 +3,13 @@ import PostFooterIconList from "@/components/molecules/PostFooterIconList/PostFo
 import LikedIconList from "@/components/molecules/LikedIconList/LikedIconList";
 import UserPostCommentSection from "@/components/molecules/UserPost/UserPostBottom/UserPostCommentSection/UserPostCommentSection";
 import useUserPostBottom from "./useUserPostBottom";
+import { fallbackUserProfileImage } from "@/constants/imageUrls";
 
 export default function UserPostBottom({
-  user_avatar,
   likesCount,
   commentsCount,
   setLikesCount,
   setCommentsCount,
-  post_id,
-  user_has_liked,
-  like_type,
-  post_user_id,
-  is_repost,
-  original_post_id,
   setPosts,
   post,
 }) {
@@ -38,18 +32,16 @@ export default function UserPostBottom({
         <hr className="border-gray-200 mx-1"></hr>
         <div className="flex justify-between ">
           <div className="flex flex-row justify-center items-center px-4 rounded-md hover:bg-slate-100 my-2">
-            <UserImage imageUrl={user_avatar} width={24} height={24} />
+            <UserImage
+              imageUrl={post.user_avatar ?? fallbackUserProfileImage}
+              width={24}
+              height={24}
+            />
           </div>
 
           <PostFooterIconList
             handleClickOnComment={handleClickOnComment}
-            post_id={post_id}
-            user_has_liked={user_has_liked}
-            like_type={like_type}
             setLikesCount={setLikesCount}
-            post_user_id={post_user_id}
-            is_repost={is_repost}
-            original_post_id={original_post_id}
             setPosts={setPosts}
             post={post}
           />
@@ -58,10 +50,10 @@ export default function UserPostBottom({
       {showCommentSection && (
         <div>
           <UserPostCommentSection
-            post_id={post_id}
+            post_id={post.id}
             setCommentsCount={setCommentsCount}
-            is_repost={is_repost}
-            original_post_id={original_post_id}
+            is_repost={post.is_repost}
+            original_post_id={post.is_repost ? post.original_post_id : null}
           />
         </div>
       )}
