+++ frontend/constants/icons.js
@@ -7,4 +7,37 @@ const icons = [
   { src: "/icons/laugh.svg", alt: "Laugh" },
 ];
 
+const iconsData = [
+  {
+    src: "/icons/like.svg",
+    width: 12,
+    height: 12,
+    alt: "Like",
+    text: "Like",
+  },
+  {
+    src: "/icons/comment.svg",
+    width: 20,
+    height: 20,
+    alt: "Comment",
+    text: "Comment",
+  },
+  {
+    src: "/icons/repost.svg",
+    width: 15,
+    height: 15,
+    alt: "Repost",
+    text: "Repost",
+  },
+  {
+    src: "/icons/send.svg",
+    width: 12,
+    height: 12,
+    alt: "Send",
+    text: "Send",
+  },
+];
+
+export { iconsData };
+
 export default icons;
